+++ UserForm.js
@@ -12,16 +12,13 @@ import RadioButton from '@folio/stripes-components/lib/RadioButton';
 import Datepicker from '@folio/stripes-components/lib/Datepicker';
 import AddressEditList from '@folio/stripes-components/lib/structures/AddressFieldGroup/AddressEdit/AddressEditList';
 import fetch from 'isomorphic-fetch';
-import { Field, reduxForm } from 'redux-form';
+import { Field } from 'redux-form';
+import stripesForm from '@folio/stripes-form';
 
 import { countriesOptions } from './data/countries';
-import { addressTypeOptions } from './data/addressTypes';
 import Autocomplete from './lib/Autocomplete';
-
-const addressFields = {
-  country: { component: Autocomplete, props: { dataOptions: countriesOptions } },
-  addressType: { component: Select, props: { dataOptions: addressTypeOptions, fullWidth: true, placeholder: 'select address type' } },
-};
+import { toListAddresses } from './converters/address';
+import { toAddressTypeOptions } from './converters/address_type';
 
 const sys = require('stripes-loader'); // eslint-disable-line
 const okapiUrl = sys.okapi.url;
@@ -96,6 +93,7 @@ class UserForm extends React.Component {
       userGroups: PropTypes.arrayOf(PropTypes.object),
       contactTypes: PropTypes.arrayOf(PropTypes.object),
     }),
+    addressTypes: PropTypes.arrayOf(PropTypes.object),
   };
 
   constructor(props) {
@@ -112,6 +110,7 @@ class UserForm extends React.Component {
       onCancel,
       initialValues,
       optionLists,
+      addressTypes,
     } = this.props;
 
     /* Menues for Add User workflow */
@@ -122,6 +121,13 @@ class UserForm extends React.Component {
       label: `${g.group} (${g.desc})`, value: g.id, selected: initialValues.patronGroup === g.id }));
     const contactTypeOptions = (optionLists.contactTypes || []).map(g => ({
       label: g.desc, value: g.id, selected: initialValues.preferredContactTypeId === g.id }));
+
+    initialValues.personal.addresses = toListAddresses(initialValues.personal.addresses, addressTypes);
+
+    const addressFields = {
+      country: { component: Autocomplete, props: { dataOptions: countriesOptions } },
+      addressType: { component: Select, props: { dataOptions: toAddressTypeOptions(addressTypes), fullWidth: true, placeholder: 'Select address type' } },
+    };
 
     return (
       <form id="form-user" style={{ height: '100%', overflow: 'auto' }}>
@@ -198,9 +204,10 @@ class UserForm extends React.Component {
   }
 }
 
-export default reduxForm({
+export default stripesForm({
   form: 'userForm',
   validate,
   asyncValidate,
   asyncBlurFields: ['username'],
+  navigationCheck: true,
 })(UserForm);