+++ Users.js
@@ -100,10 +100,11 @@ class Users extends React.Component {
       tenant: PropTypes.string.isRequired,
       token: PropTypes.string.isRequired,
     }).isRequired,
+    onSelectRow: PropTypes.func,
   };
 
   static manifest = Object.freeze({
-    addUserMode: {},
+    addUserMode: { initialValue: { mode: false } },
     userCount: { initialValue: INITIAL_RESULT_COUNT },
     users: {
       type: 'okapi',
@@ -134,15 +135,27 @@ class Users extends React.Component {
       path: 'groups',
       records: 'usergroups',
     },
+    addressTypes: {
+      type: 'okapi',
+      path: 'addresstypes',
+      records: 'addressTypes',
+    },
   });
 
   constructor(props) {
     super(props);
 
     const query = props.location.search ? queryString.parse(props.location.search) : {};
+
+    let initiallySelected = {};
+    if (/users\/view/.test(this.props.location.pathname)) {
+      const id = /view\/(.*)\//.exec(this.props.location.pathname)[1];
+      initiallySelected = { id };
+    }
+
     this.state = {
       filters: initialFilterState(filterConfig, query.filters),
-      selectedItem: {},
+      selectedItem: initiallySelected,
       searchTerm: query.query || '',
       sortOrder: query.sort || '',
     };
@@ -154,10 +167,8 @@ class Users extends React.Component {
     this.connectedViewUser = props.stripes.connect(ViewUser);
     const logger = props.stripes.logger;
     this.log = logger.log.bind(logger);
-  }
 
-  componentWillMount() {
-    if (_.isEmpty(this.props.data.addUserMode)) this.props.mutator.addUserMode.replace({ mode: false });
+    this.anchoredRowFormatter = this.anchoredRowFormatter.bind(this);
   }
 
   componentWillReceiveProps(nextProps) {
@@ -203,7 +214,7 @@ class Users extends React.Component {
     this.transitionToParams({ sort: sortOrder });
   }
 
-  onSelectRow = (e, meta) => {
+  onSelectRow = this.props.onSelectRow ? this.props.onSelectRow : (e, meta) => {
     const userId = meta.id;
     const username = meta.username;
     this.log('action', `clicked ${userId}, selected user =`, meta);
@@ -239,6 +250,10 @@ class Users extends React.Component {
     this.props.mutator.userCount.replace(this.props.data.userCount + RESULT_COUNT_INCREMENT);
   }
 
+  getRowURL(rowData) {
+    return `/users/view/${rowData.id}/${rowData.username}${this.props.location.search}`;
+  }
+
   performSearch = _.debounce((query) => {
     this.log('action', `searched for '${query}'`);
     this.transitionToParams({ query });
@@ -250,7 +265,7 @@ class Users extends React.Component {
 
   create = (data) => {
     if (data.personal.addresses) {
-      data.personal.addresses = toUserAddresses(data.personal.addresses); // eslint-disable-line no-param-reassign
+      data.personal.addresses = toUserAddresses(data.personal.addresses, this.props.data.addressTypes); // eslint-disable-line no-param-reassign
     }
 
     // extract creds object from user object
@@ -265,10 +280,11 @@ class Users extends React.Component {
 
   postCreds = (username, creds) => {
     this.log('xhr', `POST credentials for new user '${username}':`, creds);
+    const localCreds = Object.assign({}, creds, creds.password ? {} : { password: '' });
     fetch(`${this.okapi.url}/authn/credentials`, {
       method: 'POST',
       headers: Object.assign({}, { 'X-Okapi-Tenant': this.okapi.tenant, 'X-Okapi-Token': this.okapi.token, 'Content-Type': 'application/json' }),
-      body: JSON.stringify(creds),
+      body: JSON.stringify(localCreds),
     }).then((response) => {
       if (response.status >= 400) {
         this.log('xhr', 'Users. POST of creds failed.');
@@ -300,6 +316,29 @@ class Users extends React.Component {
     this.props.history.push(`${this.props.match.path}${this.props.location.search}`);
   }
 
+  // custom row formatter to wrap rows in anchor tags.
+  anchoredRowFormatter(
+    { rowIndex,
+      rowClass,
+      rowData,
+      cells,
+      rowProps,
+      labelStrings,
+    },
+  ) {
+    return (
+      <a
+        href={this.getRowURL(rowData)} key={`row-${rowIndex}`}
+        aria-label={labelStrings && labelStrings.join('...')}
+        role="listitem"
+        className={rowClass}
+        {...rowProps}
+      >
+        {cells}
+      </a>
+    );
+  }
+
   render() {
     const { data, stripes } = this.props;
     const users = data.users || [];
@@ -312,7 +351,7 @@ class Users extends React.Component {
       Active: user => user.active,
       Name: user => `${_.get(user, ['personal', 'lastName'], '')}, ${_.get(user, ['personal', 'firstName'], '')}`,
       'Patron Group': (user) => {
-        const pg = data.patronGroups.filter(g => g.id === user.patronGroup)[0];
+        const pg = this.props.data.patronGroups.filter(g => g.id === user.patronGroup)[0];
         return pg ? pg.group : '?';
       },
       'User ID': user => user.username,
@@ -323,7 +362,7 @@ class Users extends React.Component {
       this.props.stripes.hasPerm('users.item.get') ?
         (<Route
           path={`${this.props.match.path}/view/:userid/:username`}
-          render={props => <this.connectedViewUser stripes={stripes} okapi={this.okapi} paneWidth="44%" onClose={this.collapseDetails} {...props} />}
+          render={props => <this.connectedViewUser stripes={stripes} okapi={this.okapi} paneWidth="44%" onClose={this.collapseDetails} addressTypes={data.addressTypes} {...props} />}
         />) :
         (<div
           style={{
@@ -387,6 +426,8 @@ class Users extends React.Component {
             loading={resource ? resource.isPending : false}
             autosize
             virtualize
+            ariaLabel={'User search results'}
+            rowFormatter={this.anchoredRowFormatter}
           />
         </Pane>
 
@@ -395,6 +436,7 @@ class Users extends React.Component {
           <UserForm
             id="userform-adduser"
             initialValues={{ active: true, personal: { preferredContactTypeId: '002' } }}
+            addressTypes={data.addressTypes}
             onSubmit={(record) => { this.create(record); }}
             onCancel={this.onClickCloseNewUser}
             okapi={this.okapi}