+++ converters/address.js
@@ -1,42 +1,56 @@
 import _ from 'lodash';
 import { countriesByCode, countriesByName } from '../data/countries';
-import { addressTypesByDesc, addressTypesById } from '../data/addressTypes';
+import { getAddressTypesByName, getAddressTypesById } from './address_type';
+import { hashCode } from 'hashcode';
 
-function toListAddress(addr) {
+function toListAddress(addr, addrType) {
+  if (addr.id) return { ...addr };
+
   const country = (addr.countryId) ? countriesByCode[addr.countryId].country : '';
-  const addressType = (addr.addressTypeId) ? addressTypesById[addr.addressTypeId].desc : '';
+  const addressType = _.get(addrType, ['addressType'], '');
+  const id = hashCode().value(addr).toString();
 
-  return Object.assign(addr, {
-    id: addr.id || _.uniqueId(),
-    stateRegion: addr.region,
+  return {
+    id,
+    addressLine1: addr.addressLine1,
+    addressLine2: addr.addressLine2,
+    city: addr.city,
+    primaryAddress: addr.primaryAddress,
     primary: addr.primaryAddress,
+    stateRegion: addr.region,
     zipCode: addr.postalCode,
     country,
     addressType,
-  });
+  };
 }
 
-function toUserAddress(addr) {
-  const addressTypeId = (addr.addressType) ? addressTypesByDesc[addr.addressType].id : '';
+function toUserAddress(addr, addrType) {
   const countryId = (addr.country) ? countriesByName[addr.country].alpha2 : '';
-
+  const addressTypeId = _.get(addrType, ['id'], '');
   return {
     addressLine1: addr.addressLine1,
     addressLine2: addr.addressLine2,
     city: addr.city,
+    primaryAddress: addr.primaryAddress,
     region: addr.stateRegion,
     postalCode: addr.zipCode,
-    primaryAddress: addr.primary || addr.primaryAddress, // TODO: cleanup after <AddressFieldGroup> is fixed
     addressTypeId,
     countryId,
   };
 }
 
-export function toListAddresses(addresses) {
+export function toListAddresses(addresses, addressTypes) {
   if (!addresses || !addresses.length) return addresses;
-  return Array.from(addresses).map(toListAddress);
+
+  const addressTypesById = getAddressTypesById(addressTypes);
+  return addresses.map(addr =>
+    toListAddress(addr, addressTypesById[addr.addressTypeId]));
 }
 
-export function toUserAddresses(addresses) {
-  return addresses.map(toUserAddress);
+export function toUserAddresses(addresses, addressTypes) {
+  if (!addresses || !addresses.length) return addresses;
+
+  const addressTypesByName = getAddressTypesByName(addressTypes);
+  return addresses.map(addr =>
+    toUserAddress(addr, addressTypesByName[addr.addressType]));
 }