+++ util.js
@@ -5,6 +5,13 @@ export function formatDate(dateStr, locale) {
   return new Date(Date.parse(dateStr)).toLocaleDateString(locale);
 }
 
+export function futureDate(dateStr, locale, days) {
+  if (!dateStr) return dateStr;
+  const date = new Date(Date.parse(dateStr));
+  date.setDate(date.getDate() + days);
+  return date.toLocaleDateString(locale);
+}
+
 export function getFullName(user) {
   return `${_.get(user, ['personal', 'lastName'], '')},
     ${_.get(user, ['personal', 'firstName'], '')}